requirements.txt
APScheduler==3.10.1
async-timeout==4.0.2
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
Flask==2.2.5
Flask-APScheduler==1.12.4
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
nltk==3.8.1
packaging==23.1
parsedatetime==2.6
Pillow==9.5.0
psutil==5.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
regex==2023.5.5
six==1.16.0
SQLAlchemy==2.0.12
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
validators==0.20.0
Werkzeug==2.2.3
zipp==3.15.0
python-alipay-sdk==3.2.0
pyOpenSSL==23.1.1
pycryptodomex==3.17
PyCryptodome==3.17
webargs==8.2.0
jsonpickle==3.0.1
pandas==2.2.2
flask-siwadoc==0.2.2
gevent==22.10.2
flask-socketio==5.3.6
eventlet==0.36.1